+++ httprest/api.py
@@ -4,6 +4,8 @@ from typing import Optional as _Optional
 
 from httprest.http import HTTPClient as _HTTPClient
 from httprest.http import HTTPResponse as _HTTPResponse
+from httprest.http.auth import BaseAuth as _BaseAuth
+from httprest.http.cert import ClientCertificate as _ClientCertificate
 from httprest.http.urllib_client import UrllibHTTPClient as _UrllibHTTPClient
 
 
@@ -31,15 +33,25 @@ class API:
         endpoint: str,
         json: _Optional[dict] = None,
         headers: _Optional[dict] = None,
+        params: _Optional[dict] = None,
+        auth: _Optional[_BaseAuth] = None,
+        cert: _Optional[_ClientCertificate] = None,
     ) -> _HTTPResponse:
+        # pylint: disable=too-many-arguments
         """Make API request.
 
-        :param method: request HTTP method (case-insensitive)
         :param endpoint: API endpoint. Will be appended to the base URL
-        :param json: JSON data to send in the request body
+
+        Other parameters are the same as for the `HTTPClient.request` method
         """
         return self._http_client.request(
-            method, self._build_url(endpoint), json, headers
+            method=method,
+            url=self._build_url(endpoint),
+            json=json,
+            headers=headers,
+            params=params,
+            auth=auth,
+            cert=cert,
         )
 
     def _build_url(self, endpoint: str) -> str: