+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+  * `http.auth` module. Support for Basic auth
+  * `params` argument for the `HTTPClient.request` to simplify query parameters usage
+  * `cert` argument for the `HTTPClient.request` to allow client side certificates
 
 ## [0.1.2] - 2024-09-20
 ### Added